refactor: modulariza componentes do portfólio para melhor organização
Separa Header, Footer e ProjectModal em componentes individuais
para melhorar a organização do código e facilitar manutenção.

Remove imports não utilizados do Dialog e código redundante do modal.
Melhora alinhamento de texto nos estilos e corrige inconsistências
de formatação.

Facilita futuras atualizações e escalabilidade através da modularização
dos elementos da interface.
