+++ frontend/src/App.jsx
@@ -1,19 +1,32 @@
-"use client"
+"use client";
 
-import React, { useState, useRef, useEffect } from "react";
-import './styles/Header.css';
-import './styles/index.css';
-import './styles/dark-mode.css';
-import './styles/media-queries.css';
-import './styles/animations.css';
-import './styles/Popup.css';
+import React, { useState, useEffect } from "react";
+import "./styles/Header.css";
+import "./styles/index.css";
+import "./styles/dark-mode.css";
+import "./styles/media-queries.css";
+import "./styles/animations.css";
+import "./styles/Popup.css";
 import { FaBriefcase } from "react-icons/fa";
 import {
-  FaMoon, FaSun, FaGlobe, FaDownload, FaGithub, FaLinkedin, FaEnvelope, FaExternalLinkAlt, FaCode, FaDatabase, FaTools, FaCloud, FaDesktop, FaCogs, FaRocket, FaUserAstronaut, FaPhone, FaShieldAlt, FaUsers,
+  FaGithub,
+  FaLinkedin,
+  FaEnvelope,
+  FaExternalLinkAlt,
+  FaCode,
+  FaDatabase,
+  FaTools,
+  FaCloud,
+  FaDesktop,
+  FaCogs,
+  FaRocket,
+  FaUserAstronaut,
+  FaPhone,
+  FaShieldAlt,
 } from "react-icons/fa";
-import {
-  Dialog, DialogContent, DialogDescription, DialogTrigger,
-} from "./components/ui/dialog"
+import Header from "./components/Header";
+import Footer from "./components/Footer";
+import ProjectModal from "./components/ProjectModal";
 
 import { SiJira } from "react-icons/si";
 
@@ -23,13 +36,12 @@ function App() {
   const [selectedProject, setSelectedProject] = useState(null);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [isNavOpen, setIsNavOpen] = useState(false);
-  const modalRef = useRef(null);
 
-  const handleLang = () => setLang(lang === "pt-br" ? "en" : "pt-br")
+  const handleLang = () => setLang(lang === "pt-br" ? "en" : "pt-br");
   const toggleDarkMode = () => {
-    setDarkMode(!darkMode)
-    document.body.classList.toggle("dark-mode")
-  }
+    setDarkMode(!darkMode);
+    document.body.classList.toggle("dark-mode");
+  };
 
   const toggleNav = () => {
     setIsNavOpen(!isNavOpen);
@@ -37,275 +49,6 @@ function App() {
 
   const closeNav = () => {
     setIsNavOpen(false);
-  };
-  
-  const handleImageClick = (imageSrc, imageTitle) => {
-    // Calcular as dimensões da tela
-    const screenWidth = window.screen.availWidth;
-    const screenHeight = window.screen.availHeight;
-    
-    // Definir dimensões da janela (responsivo)
-    const windowWidth = Math.min(1400, screenWidth * 0.9);
-    const windowHeight = Math.min(900, screenHeight * 0.9);
-    
-    // Calcular posição centralizada
-    const left = (screenWidth - windowWidth) / 2;
-    const top = (screenHeight - windowHeight) / 2;
-    
-    // CSS organizado do iframe.css
-    const iframeCss = `
-      * {
-        margin: 0;
-        padding: 0;
-        box-sizing: border-box;
-      }
-  
-      html, body {
-        width: 100%;
-        height: 100%;
-        overflow: hidden;
-        background: linear-gradient(135deg, #f0f8ff 0%, #e3f2fd 100%);
-      }
-  
-      .image-viewer-container {
-        width: 100vw;
-        height: 100vh;
-        display: flex;
-        flex-direction: column;
-        position: relative;
-      }
-  
-      .image-viewer-header {
-        background: linear-gradient(135deg, #1565c0 0%, #42a5f5 100%);
-        color: white;
-        padding: 1rem 2rem;
-        text-align: center;
-        box-shadow: 0 4px 20px rgba(21, 101, 192, 0.3);
-        z-index: 10;
-        position: relative;
-      }
-  
-      .image-viewer-header h1 {
-        margin: 0;
-        font-size: 1.8rem;
-        font-weight: 700;
-        text-shadow: 0 2px 8px rgba(0,0,0,0.2);
-      }
-  
-      .image-viewer-body {
-        flex: 1;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding: 20px;
-        background: linear-gradient(135deg, #4a6b8a 0%, #5c7a9a 50%, #607fa9 100%);
-        position: relative;
-        overflow: hidden;
-      }
-  
-      .image-viewer-wrapper {
-        max-width: 90%;
-        max-height: 100%;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        border-radius: 12px;
-        box-shadow: 0 20px 60px rgba(30, 136, 229, 0.25);
-        background: linear-gradient(135deg, #607fa9 0%, #395e7a 50%, #000102 100%);
-        padding: 15px;
-      }
-  
-      .image-viewer-img {
-        max-width: 100%;
-        max-height: 45rem;
-        width: auto;
-        height: auto;
-        border-radius: 8px;
-        object-fit: contain;
-        transition: all 0.8s cubic-bezier(0.45, 0.56, 0.55, 0.99);
-        display: block;
-        cursor: pointer;
-        transform-origin: center center;
-        filter: brightness(1) contrast(1) saturate(1);
-      }
-
-      .image-viewer-close-btn {
-        position: fixed;
-        top: 20px;
-        right: 20px;
-        background: linear-gradient(135deg, #e53e3e 0%, #c53030 100%);
-        border: none;
-        border-radius: 50%;
-        width: 50px;
-        height: 50px;
-        color: white;
-        font-size: 1.6rem;
-        cursor: pointer;
-        box-shadow: 0 8px 25px rgba(229, 62, 62, 0.4);
-        transition: all 0.3s ease;
-        z-index: 1000;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        font-weight: 700;
-      }
-  
-      .image-viewer-close-btn:hover {
-        background: linear-gradient(135deg, #c53030 0%, #9b2c2c 100%);
-        transform: scale(1.1);
-        box-shadow: 0 12px 35px rgba(229, 62, 62, 0.5);
-      }
-  
-      .image-viewer-close-btn:active {
-        transform: scale(0.95);
-      }
-  
-      .image-viewer-loading {
-        color: #1565c0;
-        font-size: 18px;
-        text-align: center;
-        padding: 3rem;
-        font-weight: 500;
-      }
-  
-      .image-viewer-footer-info {
-        position: absolute;
-        bottom: 15px;
-        left: 50%;
-        transform: translateX(-50%);
-        color: #1565c0;
-        font-size: 0.85rem;
-        text-align: center;
-        opacity: 0.8;
-        background: rgba(255, 255, 255, 0.9);
-        padding: 0.5rem 1rem;
-        border-radius: 20px;
-        box-shadow: 0 2px 10px rgba(30, 136, 229, 0.1);
-      }
-  
-      @media (min-width: 1200px) {
-        .image-viewer-header h1 {
-          font-size: 2.2rem;
-        }
-        .image-viewer-body {
-          padding: 30px;
-        }
-        .image-viewer-wrapper {
-          padding: 25px;
-        }
-      }
-  
-      @media (max-width: 768px) {
-        .image-viewer-header {
-          padding: 0.8rem 1.5rem;
-        }
-        .image-viewer-header h1 {
-          font-size: 1.4rem;
-        }
-        .image-viewer-body {
-          padding: 15px;
-        }
-        .image-viewer-close-btn {
-          width: 45px;
-          height: 45px;
-          font-size: 1.4rem;
-        }
-      }
-  
-      :fullscreen .image-viewer-container, :-webkit-full-screen .image-viewer-container, :-moz-full-screen .image-viewer-container {
-        width: 100vw;
-        height: 100vh;
-      }
-    `;
-    
-    // Criar uma nova janela centralizada
-    const newWindow = window.open(
-      '', 
-      '_blank', 
-      `width=${windowWidth},height=${windowHeight},left=${left},top=${top},scrollbars=no,resizable=yes,toolbar=no,menubar=no,location=no,status=no`
-    );
-    
-    if (newWindow) {
-      newWindow.document.write(`
-        <!DOCTYPE html>
-        <html>
-        <head>
-          <meta charset="utf-8">
-          <meta name="viewport" content="width=device-width, initial-scale=1.0">
-          <title>${imageTitle}</title>
-          <link rel="preconnect" href="https://fonts.googleapis.com">
-          <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-          <link href="https://fonts.googleapis.com/css2?family=Segoe+UI:wght@400;500;600;700&display=swap" rel="stylesheet">
-          <style>
-            ${iframeCss}
-          </style>
-        </head>
-        <body>
-          <div class="image-viewer-container">
-            <button class="image-viewer-close-btn" onclick="window.close()" title="Fechar janela">&times;</button>
-            
-            <div class="image-viewer-header">
-              <h1>${imageTitle}</h1>
-            </div>
-            
-            <div class="image-viewer-body">
-              <div class="image-viewer-wrapper">
-                <img src="${imageSrc}" alt="${imageTitle}" class="image-viewer-img"
-                     onerror="this.parentElement.innerHTML='<div class=image-viewer-loading>❌ Erro ao carregar imagem</div>'"
-                     onload="this.style.opacity=1" style="opacity:0;transition:opacity 0.5s">
-              </div>
-              
-              <div class="image-viewer-footer-info">
-                ESC para fechar • F11 para tela cheia
-              </div>
-            </div>
-          </div>
-          
-          <script>
-            window.focus();
-            
-            document.addEventListener('keydown', function(e) {
-              if (e.key === 'Escape') {
-                window.close();
-              }
-              if (e.key === 'F11') {
-                e.preventDefault();
-                if (!document.fullscreenElement) {
-                  document.documentElement.requestFullscreen();
-                } else {
-                  document.exitFullscreen();
-                }
-              }
-            });
-            
-            window.addEventListener('resize', function() {
-              if (!document.fullscreenElement) {
-                const screenWidth = window.screen.availWidth;
-                const screenHeight = window.screen.availHeight;
-                const windowWidth = window.outerWidth;
-                const windowHeight = window.outerHeight;
-                const left = (screenWidth - windowWidth) / 2;
-                const top = (screenHeight - windowHeight) / 2;
-                window.moveTo(left, top);
-              }
-            });
-            
-            document.addEventListener('fullscreenchange', function() {
-              const container = document.querySelector('.image-viewer-container');
-              if (document.fullscreenElement) {
-                container.style.width = '100vw';
-                container.style.height = '100vh';
-              }
-            });
-          </script>
-        </body>
-        </html>
-      `);
-      newWindow.document.close();
-      newWindow.focus();
-    } else {
-      alert('⚠️ Popups estão bloqueados!\n\nPor favor, permita popups para este site e tente novamente.');
-    }
   };
 
   useEffect(() => {
@@ -339,42 +82,47 @@ function App() {
         {
           imagem: "/images/homeFlowtime.png",
           titulo: "Dashboard Interativo",
-          descricao: "Visualize rapidamente os principais indicadores da clínica, como agendamentos, atendimentos e receitas. Desenvolvimento completo do frontend e implementação dos gráficos interativos."
+          descricao:
+            "Visualize rapidamente os principais indicadores da clínica, como agendamentos, atendimentos e receitas. Desenvolvimento completo do frontend e implementação dos gráficos interativos.",
         },
         {
           imagem: "/images/AppointmentsFlowTime.png",
           titulo: "Gestão de Agendas",
-          descricao: "Permite marcar, editar e visualizar consultas de forma simples e rápida. Interface responsiva e intuitiva desenvolvida integralmente."
+          descricao:
+            "Permite marcar, editar e visualizar consultas de forma simples e rápida. Interface responsiva e intuitiva desenvolvida integralmente.",
         },
         {
           imagem: "/images/FichaAnamnese.png",
           titulo: "Relatórios em PDF",
-          descricao: "Gere relatórios detalhados de atendimentos, receitas e pacientes em poucos cliques. Frontend da funcionalidade e sistema de tokens de segurança para acesso."
-        }
+          descricao:
+            "Gere relatórios detalhados de atendimentos, receitas e pacientes em poucos cliques. Frontend da funcionalidade e sistema de tokens de segurança para acesso.",
+        },
       ],
       minhasResponsabilidades: [
         "🎨 Frontend Completo - Desenvolvimento de toda interface do usuário",
         "📊 Gráficos Interativos - Implementação de dashboards e visualizações",
         "🔐 Sistema de Tokens - Desenvolvimento de autenticação e segurança",
         "📱 Responsividade - Adaptação para dispositivos móveis e desktop",
-        "✨ UX/UI - Implementação do design e experiência do usuário"
+        "✨ UX/UI - Implementação do design e experiência do usuário",
       ],
       equipe: [
         {
           nome: "Eduardo Wagner",
           papel: "Frontend Developer",
-          responsabilidades: "Frontend completo, gráficos interativos, sistema de tokens de segurança"
+          responsabilidades:
+            "Frontend completo, gráficos interativos, sistema de tokens de segurança",
         },
         {
-          nome: "Victor Bueno", 
+          nome: "Victor Bueno",
           papel: "Backend Developer",
-          responsabilidades: "Toda arquitetura do backend, APIs, banco de dados"
+          responsabilidades:
+            "Toda arquitetura do backend, APIs, banco de dados",
         },
         {
           nome: "Lucas Ulbrich",
           papel: "UI/UX Designer",
-          responsabilidades: "Prototipagem das páginas no Figma, design system"
-        }
+          responsabilidades: "Prototipagem das páginas no Figma, design system",
+        },
       ],
       frontend: [
         "HTML5",
@@ -383,7 +131,7 @@ function App() {
         "JavaScript",
         "Electron.js",
         "SweetAlert2",
-        "Chart.js"
+        "Chart.js",
       ],
       backend: [
         "Node.js",
@@ -395,14 +143,14 @@ function App() {
         "node-cache",
         "web-push",
         "path",
-        "cross-env"
+        "cross-env",
       ],
       database: [
         "Firebase Firestore",
         "Firebase Database",
         "Firebase Storage",
         "Firebase Authentication",
-        "Firebase Admin SDK"
+        "Firebase Admin SDK",
       ],
       libraries: [
         "browser-image-compression",
@@ -412,7 +160,7 @@ function App() {
         "SweetAlert2",
         "Choices.js",
         "html2pdf.js",
-        "SignaturePad.js"
+        "SignaturePad.js",
       ],
       integrations: [
         "API IBGE",
@@ -420,11 +168,11 @@ function App() {
         "Web Push API",
         "Firebase Auth API",
         "Service Workers",
-        "Local Storage"
+        "Local Storage",
       ],
       status: "Finalizado",
       image: "/images/FlowTime.png",
-      link: "#"
+      link: "#",
     },
     {
       title: "TonnerTrack",
@@ -437,18 +185,21 @@ function App() {
         {
           imagem: "/images/GerenciamentoTonnerTrack.png",
           titulo: "Controle de Suprimentos",
-          descricao: "Gerencie o estoque de toners e impressoras, evitando desperdícios e falta de material. Sistema completo desenvolvido do zero."
+          descricao:
+            "Gerencie o estoque de toners e impressoras, evitando desperdícios e falta de material. Sistema completo desenvolvido do zero.",
         },
         {
           imagem: "/images/Relatórios.png",
           titulo: "Relatórios Detalhados",
-          descricao: "Gere relatórios completos sobre o uso de impressoras, facilitando a tomada de decisão. Interface e backend desenvolvidos integralmente."
+          descricao:
+            "Gere relatórios completos sobre o uso de impressoras, facilitando a tomada de decisão. Interface e backend desenvolvidos integralmente.",
         },
         {
           imagem: "/images/NotificacaoTonnerTrac.png",
           titulo: "Alertas Inteligentes",
-          descricao: "Receba notificações quando o estoque estiver baixo ou houver necessidade de manutenção. Sistema de notificações desenvolvido por completo."
-        }
+          descricao:
+            "Receba notificações quando o estoque estiver baixo ou houver necessidade de manutenção. Sistema de notificações desenvolvido por completo.",
+        },
       ],
       minhasResponsabilidades: [
         "🎨 Desenvolvimento Completo - Frontend e Backend do zero",
@@ -456,14 +207,15 @@ function App() {
         "🔐 Autenticação - Sistema de login e controle de acesso",
         "📊 Relatórios - Geração de relatórios e dashboards",
         "📱 Interface Responsiva - Design adaptável para todos dispositivos",
-        "🔧 Deploy e Manutenção - Implementação em ambiente de produção"
+        "🔧 Deploy e Manutenção - Implementação em ambiente de produção",
       ],
       equipe: [
         {
           nome: "Eduardo Wagner",
           papel: "Desenvolvedor Full Stack Solo",
-          responsabilidades: "Desenvolvimento completo do sistema, desde planejamento até implementação final"
-        }
+          responsabilidades:
+            "Desenvolvimento completo do sistema, desde planejamento até implementação final",
+        },
       ],
       frontend: [
         "HTML5",
@@ -474,7 +226,7 @@ function App() {
         "Bootstrap Icons",
         "Chart.js",
         "SweetAlert2",
-        "Electron.js"
+        "Electron.js",
       ],
       backend: [
         "Node.js",
@@ -484,26 +236,18 @@ function App() {
         "Multer",
         "Socket.io",
         "dotenv",
-        "cookie-parser"
+        "cookie-parser",
       ],
-      database: [
-        "PostgreSQL",
-        "pg"
-      ],
-      libraries: [
-        "firebase",
-        "firebase-admin",
-        "path",
-        "electron-updater"
-      ],
+      database: ["PostgreSQL", "pg"],
+      libraries: ["firebase", "firebase-admin", "path", "electron-updater"],
       integrations: [
         "Firebase Authentication",
         "Firebase Cloud Storage",
-        "API RESTful própria"
+        "API RESTful própria",
       ],
       status: "Finalizado",
       image: "/images/TonnerTrack.png",
-      link: "#"
+      link: "#",
     },
     {
       title: "ThermalTech",
@@ -516,46 +260,51 @@ function App() {
         {
           imagem: "/images/Monitoriamento.png",
           titulo: "Monitoramento em Tempo Real",
-          descricao: "Acompanhe a temperatura e o status dos equipamentos em tempo real através de sensores IoT. Frontend completo e sistema de segurança implementado."
+          descricao:
+            "Acompanhe a temperatura e o status dos equipamentos em tempo real através de sensores IoT. Frontend completo e sistema de segurança implementado.",
         },
         {
           imagem: "/images/ConsultarChamados.png",
           titulo: "Gestão de Chamados",
-          descricao: "Registre, acompanhe e resolva chamados técnicos de forma centralizada e eficiente. Interface desenvolvida com sistema de auditoria avançado."
+          descricao:
+            "Registre, acompanhe e resolva chamados técnicos de forma centralizada e eficiente. Interface desenvolvida com sistema de auditoria avançado.",
         },
         {
           imagem: "/images/ChatTT.png",
           titulo: "Chat Integrado",
-          descricao: "Comunique-se com a equipe diretamente pelo sistema, agilizando o suporte e a manutenção. Frontend e middleware de segurança implementados."
-        }
+          descricao:
+            "Comunique-se com a equipe diretamente pelo sistema, agilizando o suporte e a manutenção. Frontend e middleware de segurança implementados.",
+        },
       ],
       minhasResponsabilidades: [
         "🎨 Frontend Completo - Desenvolvimento de toda interface do sistema",
         "🛡️ Middleware de Auditoria - Sistema avançado de detecção de ameaças",
         "🔍 Detecção de Ameaças - SQL injection, XSS, bots maliciosos",
-        "📊 Threat Scoring - Sistema inteligente de pontuação de ameaças", 
+        "📊 Threat Scoring - Sistema inteligente de pontuação de ameaças",
         "🚫 Bloqueio Inteligente - IPs suspeitos e rate limiting",
         "📧 Alertas em Tempo Real - Notificações por email automáticas",
-        "📋 Relatórios de Segurança - Monitoramento contínuo automatizado"
+        "📋 Relatórios de Segurança - Monitoramento contínuo automatizado",
       ],
       equipe: [
         {
           nome: "Eduardo Wagner",
-          papel: "Frontend & Security Developer", 
-          responsabilidades: "Frontend completo, sistema de segurança avançado, middleware de auditoria, detecção de ameaças"
+          papel: "Frontend & Security Developer",
+          responsabilidades:
+            "Frontend completo, sistema de segurança avançado, middleware de auditoria, detecção de ameaças",
         },
         {
           nome: "Victor Bueno",
           papel: "Backend & IoT Developer",
-          responsabilidades: "Backend, integração IoT, sensores ESP, Firebase Realtime Database, comunicação WebSocket/SSE"
-        }
+          responsabilidades:
+            "Backend, integração IoT, sensores ESP, Firebase Realtime Database, comunicação WebSocket/SSE",
+        },
       ],
       integracaoIoT: [
         "🌡️ Captura de Temperatura - Sensores via HTTP POST no endpoint /api/sensor-data",
-        "📡 Comunicação Bidirecional - ESP via Firebase Realtime Database", 
+        "📡 Comunicação Bidirecional - ESP via Firebase Realtime Database",
         "⚡ Tempo Real - Server-Sent Events (SSE) e WebSocket",
         "🎯 Controle Remoto - Comandos de temperatura, velocidade e modo",
-        "⚠️ Alertas Automáticos - Verificação de limites por sala"
+        "⚠️ Alertas Automáticos - Verificação de limites por sala",
       ],
       frontend: [
         "HTML5",
@@ -568,7 +317,7 @@ function App() {
         "HTML2Canvas",
         "JavaScript",
         "Socket.IO",
-        "WebSocket"
+        "WebSocket",
       ],
       backend: [
         "Node.js",
@@ -577,13 +326,13 @@ function App() {
         "Socket.IO",
         "EJS",
         "MQTT",
-        "Firebase Admin SDK"
+        "Firebase Admin SDK",
       ],
       database: [
         "PostgreSQL",
         "Firebase Authentication",
         "Firebase Storage",
-        "Firebase Database"
+        "Firebase Database",
       ],
       libraries: [
         "SweetAlert2",
@@ -594,7 +343,7 @@ function App() {
         "Socket.IO",
         "Firebase",
         "MQTT",
-        "EJS"
+        "EJS",
       ],
       integrations: [
         "Firebase Authentication",
@@ -602,241 +351,62 @@ function App() {
         "Firebase Database",
         "Socket.IO",
         "MQTT",
-        "APIs REST próprias"
+        "APIs REST próprias",
       ],
       status: "Em Desenvolvimento",
       image: "/images/ThermalTech.png",
-      link: "#"
-    }
-  ]
+      link: "#",
+    },
+  ];
 
   const skills = {
-    frontend: ["HTML5", "CSS3", "JavaScript ES6+", "Bootstrap", "EJS", "Chart.js", "jQuery"],
+    frontend: [
+      "HTML5",
+      "CSS3",
+      "JavaScript ES6+",
+      "Bootstrap",
+      "EJS",
+      "Chart.js",
+      "jQuery",
+    ],
     backend: ["Node.js", "Express.js", "Python"],
-    database: ["PostgreSQL", "Firebase", "Firestore", "Firebase Storage", "Firebase Auth"],
+    database: [
+      "PostgreSQL",
+      "Firebase",
+      "Firestore",
+      "Firebase Storage",
+      "Firebase Auth",
+    ],
     cloud: ["Google Cloud Platform", "Firebase"],
     desktop: ["Electron.js", "PWA"],
     tools: ["Git", "GitHub", "SweetAlert2", "Multer", "WebSockets", "MQTT"],
-    methodologies: ["Scrum", "Jira", "Metodologias Ágeis", "Arquitetura MVC", "Trabalho em Equipe"],
+    methodologies: [
+      "Scrum",
+      "Jira",
+      "Metodologias Ágeis",
+      "Arquitetura MVC",
+      "Trabalho em Equipe",
+    ],
     learning: ["ReactJS", "React Native"],
-  }
+  };
 
   const handleProjectClick = (project) => {
     setSelectedProject(project);
     setIsModalOpen(true);
   };
 
-  <div
-    className="modal fade"
-    id="projectModal"
-    tabIndex="-1"
-    aria-labelledby="projectModalLabel"
-    aria-hidden="true"
-    ref={modalRef}
-  >
-    <div className="modal-dialog modal-lg modal-dialog-centered">
-      <div className="modal-content">
-        {selectedProject && (
-          <>
-            <div className="modal-header">
-              <h5 className="modal-title" id="projectModalLabel">
-                {selectedProject.title}
-              </h5>
-              <button
-                type="button"
-                className="modal-close-btn"
-                aria-label="Fechar"
-                onClick={() => setIsModalOpen(false)}
-              >
-                ×
-              </button>
-            </div>
-            <div className="modal-body">
-              <img
-                src={selectedProject.image || "/placeholder.svg"}
-                alt={selectedProject.title}
-                className="img-fluid rounded mb-3"
-              />
-              <p>{selectedProject.fullDescription}</p>
-              {/* Adapte aqui para mostrar tecnologias, etc. */}
-            </div>
-          </>
-        )}
-      </div>
-    </div>
-  </div>
-    
   return (
     <div className={`portfolio ${darkMode ? "dark-mode" : ""}`}>
       {/* Header estilo moderno */}
-      <nav className="navbar navbar-expand-lg navbar-light bg-white shadow-sm py-3 fixed-top">
-        <div className="container">
-          <a
-            className="navbar-brand logo-gradient"
-            href="#home"
-            onClick={closeNav}
-          >
-            🚀 EW
-          </a>
-          <button
-            className="navbar-toggler"
-            type="button"
-            onClick={toggleNav}
-            aria-controls="navbarNav"
-            aria-expanded={isNavOpen}
-            aria-label="Toggle navigation"
-          >
-            <span className="navbar-toggler-icon"></span>
-          </button>
-          <div
-            className={`collapse navbar-collapse ${isNavOpen ? "show" : ""}`}
-            id="navbarNav"
-          >
-            {/* Botão X - apenas mobile */}
-            <button
-              className="mobile-close-btn d-md-none"
-              onClick={closeNav}
-              aria-label="Fechar menu"
-            >
-              ×
-            </button>
-
-            <ul className="navbar-nav mx-auto mb-2 mb-lg-0 gap-2">
-              <li className="nav-item">
-                <a className="nav-link" href="#home" onClick={closeNav}>
-                  Home
-                </a>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="#about" onClick={closeNav}>
-                  Sobre
-                </a>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="#skills" onClick={closeNav}>
-                  Habilidades
-                </a>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="#projects" onClick={closeNav}>
-                  Projetos
-                </a>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="#experience" onClick={closeNav}>
-                  Experiência
-                </a>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="#contact" onClick={closeNav}>
-                  Contato
-                </a>
-              </li>
-            </ul>
-
-            <div className="d-flex align-items-center gap-2">
-              {/* Container para botões pequenos - lado a lado apenas no mobile */}
-              <div className="mobile-small-buttons d-md-none">
-                <button
-                  className="btn btn-outline-primary position-relative"
-                  onClick={toggleDarkMode}
-                  onMouseEnter={(e) =>
-                    e.currentTarget.classList.add("show-tooltip")
-                  }
-                  onMouseLeave={(e) =>
-                    e.currentTarget.classList.remove("show-tooltip")
-                  }
-                >
-                  {darkMode ? <FaSun /> : <FaMoon />}
-                  <span className="custom-tooltip">
-                    {darkMode ? "Modo claro" : "Modo escuro"}
-                  </span>
-                </button>
-
-                <button
-                  className="btn btn-outline-primary position-relative"
-                  onClick={handleLang}
-                  onMouseEnter={(e) =>
-                    e.currentTarget.classList.add("show-tooltip")
-                  }
-                  onMouseLeave={(e) =>
-                    e.currentTarget.classList.remove("show-tooltip")
-                  }
-                >
-                  <FaGlobe /> {lang === "pt-br" ? "EN" : "PT"}
-                  <span className="custom-tooltip">Mudar idioma</span>
-                </button>
-              </div>
-
-              {/* Versão desktop - todos os botões na mesma linha */}
-              <div className="desktop-buttons d-none d-md-flex align-items-center gap-2">
-                <button
-                  className="btn btn-outline-primary position-relative"
-                  onClick={toggleDarkMode}
-                  onMouseEnter={(e) =>
-                    e.currentTarget.classList.add("show-tooltip")
-                  }
-                  onMouseLeave={(e) =>
-                    e.currentTarget.classList.remove("show-tooltip")
-                  }
-                >
-                  {darkMode ? <FaSun /> : <FaMoon />}
-                  <span className="custom-tooltip">
-                    {darkMode ? "Modo claro" : "Modo escuro"}
-                  </span>
-                </button>
-
-                <a
-                  className="btn btn-primary d-flex align-items-center gap-2 position-relative"
-                  href="/cv.pdf"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  onMouseEnter={(e) =>
-                    e.currentTarget.classList.add("show-tooltip")
-                  }
-                  onMouseLeave={(e) =>
-                    e.currentTarget.classList.remove("show-tooltip")
-                  }
-                >
-                  <FaDownload /> Download CV
-                  <span className="custom-tooltip">Baixar currículo</span>
-                </a>
-
-                <button
-                  className="btn btn-outline-primary position-relative"
-                  onClick={handleLang}
-                  onMouseEnter={(e) =>
-                    e.currentTarget.classList.add("show-tooltip")
-                  }
-                  onMouseLeave={(e) =>
-                    e.currentTarget.classList.remove("show-tooltip")
-                  }
-                >
-                  <FaGlobe /> {lang === "pt-br" ? "EN" : "PT"}
-                  <span className="custom-tooltip">Mudar idioma</span>
-                </button>
-              </div>
-
-              {/* Botão Download CV - embaixo no mobile */}
-              <a
-                className="btn btn-primary d-flex d-md-none align-items-center gap-2 position-relative mobile-download-btn"
-                href="/cv.pdf"
-                target="_blank"
-                rel="noopener noreferrer"
-                onMouseEnter={(e) =>
-                  e.currentTarget.classList.add("show-tooltip")
-                }
-                onMouseLeave={(e) =>
-                  e.currentTarget.classList.remove("show-tooltip")
-                }
-              >
-                <FaDownload /> Download CV
-                <span className="custom-tooltip">Baixar currículo</span>
-              </a>
-            </div>
-          </div>
-        </div>
-      </nav>
+      <Header
+        darkMode={darkMode}
+        toggleDarkMode={toggleDarkMode}
+        lang={lang}
+        handleLang={handleLang}
+        isNavOpen={isNavOpen}
+        toggleNav={toggleNav}
+        closeNav={closeNav}
+      />
 
       {/* 1. Home Section */}
       <section id="home" className="hero-section">
@@ -1315,15 +885,13 @@ function App() {
                         alt={project.title}
                       />
                       <div className="project-overlay">
-                        <DialogTrigger asChild>
-                          <button
-                            className="btn btn-primary btn-sm"
-                            onClick={() => handleProjectClick(project)}
-                          >
-                            <FaExternalLinkAlt className="me-1" />
-                            Ver Detalhes
-                          </button>
-                        </DialogTrigger>
+                        <button
+                          className="btn btn-primary btn-sm"
+                          onClick={() => handleProjectClick(project)}
+                        >
+                          <FaExternalLinkAlt className="me-1" />
+                          Ver Detalhes
+                        </button>
                       </div>
                     </div>
                     <div className="project-content">
@@ -1349,234 +917,6 @@ function App() {
           </div>
         </div>
       </section>
-
-      {/* Project Details Modal */}
-      <Dialog open={isModalOpen} onOpenChange={setIsModalOpen}>
-        <DialogContent className="project-modal-content">
-          <button
-            type="button"
-            className="modal-close-btn"
-            aria-label="Fechar"
-            onClick={() => setIsModalOpen(false)}
-          >
-            &times;
-          </button>
-          {selectedProject && (
-            <>
-              <div className="project-modal-header">
-                <h2 className="project-modal-title">{selectedProject.title}</h2>
-                {selectedProject.subtitulo && (
-                  <div className="project-status-badge">
-                    {selectedProject.subtitulo}
-                  </div>
-                )}
-              </div>
-
-              <DialogDescription>
-                <div className="project-modal-main-row">
-                  <img
-                    src={selectedProject.image || "/placeholder.svg"}
-                    alt={selectedProject.title}
-                    className="project-modal-image"
-                  />
-                  <div className="project-modal-summary">
-                    {selectedProject.historia}
-                  </div>
-                </div>
-              </DialogDescription>
-
-              {/* Minhas Responsabilidades */}
-              {selectedProject.minhasResponsabilidades && (
-                <div className="project-responsibilities">
-                  <h4 className="project-responsibilities-title">
-                    🎯 Minhas Responsabilidades no Projeto
-                  </h4>
-                  <div className="project-responsibilities-container">
-                    <div className="project-responsibilities-bg-decoration"></div>
-                    <div className="project-responsibilities-content">
-                      {selectedProject.minhasResponsabilidades.map(
-                        (resp, idx) => (
-                          <div
-                            key={idx}
-                            className="project-responsibility-item"
-                          >
-                            <span className="project-responsibility-emoji">
-                              ✨
-                            </span>
-                            {resp}
-                          </div>
-                        )
-                      )}
-                    </div>
-                  </div>
-                </div>
-              )}
-
-              {/* Equipe do Projeto */}
-              {selectedProject.equipe && (
-                <div className="project-team">
-                  <h4 className="project-team-title">
-                    <FaUsers
-                      style={{ marginRight: "0.5rem", color: "#1e88e5" }}
-                    />
-                    Equipe do Projeto
-                  </h4>
-                  <div className="project-team-grid">
-                    {selectedProject.equipe.map((membro, idx) => {
-                      // Definir redes sociais para cada membro
-                      const redesSociais = {
-                        "Eduardo Wagner": {
-                          github: "https://github.com/EduardoWagner03",
-                          linkedin:
-                            "https://www.linkedin.com/in/eduardowagner03/",
-                        },
-                        "Victor Bueno": {
-                          github: "https://github.com/victorbueno920",
-                          linkedin:
-                            "https://www.linkedin.com/in/victor-bueno-365461288/",
-                        },
-                        "Lucas Ulbrich": {
-                          github: "https://github.com/lucasulbrich",
-                          linkedin:
-                            "https://www.linkedin.com/in/lucas-ulbrich/",
-                        },
-                      };
-
-                      const redes = redesSociais[membro.nome] || {
-                        github: "#",
-                        linkedin: "#",
-                      };
-
-                      return (
-                        <div key={idx} className="project-team-member">
-                          {/* Avatar placeholder */}
-                          <div className="project-team-avatar">
-                            {membro.nome
-                              .split(" ")
-                              .map((n) => n[0])
-                              .join("")}
-                          </div>
-
-                          <h6 className="project-team-member-name">
-                            {membro.nome}
-                          </h6>
-
-                          <div className="project-team-member-role">
-                            {membro.papel}
-                          </div>
-
-                          <div className="project-team-member-responsibilities">
-                            {membro.responsabilidades}
-                          </div>
-
-                          {/* Redes Sociais */}
-                          <div className="project-team-social-links">
-                            <a
-                              href={redes.github}
-                              target="_blank"
-                              rel="noopener noreferrer"
-                              className="project-team-social-link project-team-github"
-                            >
-                              <FaGithub />
-                            </a>
-
-                            <a
-                              href={redes.linkedin}
-                              target="_blank"
-                              rel="noopener noreferrer"
-                              className="project-team-social-link project-team-linkedin"
-                            >
-                              <FaLinkedin />
-                            </a>
-                          </div>
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-
-              {/* Integração IoT (só para ThermalTech) */}
-              {selectedProject.integracaoIoT && (
-                <div className="iot-integration">
-                  <h4 className="iot-integration-title">
-                    🌐 Integração IoT Desenvolvida pelo Victor
-                  </h4>
-                  <div className="iot-integration-container">
-                    <div className="iot-integration-bg-decoration"></div>
-                    <div className="iot-integration-content">
-                      {selectedProject.integracaoIoT.map((feature, idx) => (
-                        <div key={idx} className="iot-integration-item">
-                          <span className="iot-integration-emoji">🔧</span>
-                          {feature}
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                </div>
-              )}
-
-              {/* Funcionalidades do Projeto */}
-              <div className="project-features">
-                {selectedProject.funcionalidades &&
-                  selectedProject.funcionalidades.map((func, idx) => (
-                    <div key={idx} className="feature-row">
-                      <img
-                        src={func.imagem}
-                        alt={func.titulo}
-                        className="feature-img"
-                        onClick={() =>
-                          handleImageClick(func.imagem, func.titulo)
-                        }
-                      />
-                      <div className="feature-content">
-                        <h6 className="feature-title">{func.titulo}</h6>
-                        <p className="feature-desc">{func.descricao}</p>
-                      </div>
-                    </div>
-                  ))}
-              </div>
-
-              {/* Cards de tecnologias */}
-              <div className="tech-details-grid">
-                {(() => {
-                  const categories = [
-                    { key: "frontend", label: "Frontend", tag: "frontend-tag" },
-                    { key: "backend", label: "Backend", tag: "backend-tag" },
-                    { key: "database", label: "Database", tag: "database-tag" },
-                    {
-                      key: "libraries",
-                      label: "Bibliotecas",
-                      tag: "library-tag",
-                    },
-                    {
-                      key: "integrations",
-                      label: "APIs e Integrações",
-                      tag: "general-tag",
-                    },
-                  ];
-                  return categories.map(({ key, label, tag }) =>
-                    selectedProject[key] && selectedProject[key].length > 0 ? (
-                      <div className="tech-details-card" key={key}>
-                        <h6 className="tech-category-title">{label}</h6>
-                        <div className="tech-tags">
-                          {selectedProject[key].map((tech, idx) => (
-                            <span key={idx} className={`tech-tag ${tag}`}>
-                              {tech}
-                            </span>
-                          ))}
-                        </div>
-                      </div>
-                    ) : (
-                      <div className="tech-details-card empty-card" key={key} />
-                    )
-                  );
-                })()}
-              </div>
-            </>
-          )}
-        </DialogContent>
-      </Dialog>
 
       {/* 5. Experiência */}
       <section id="experience" className="experience-section section-gradient">
@@ -2223,63 +1563,18 @@ function App() {
         </div>
       </section>
 
+      {/* Modal / Iframe: */}
+      <ProjectModal
+        isModalOpen={isModalOpen}
+        setIsModalOpen={setIsModalOpen}
+        selectedProject={selectedProject}
+        handleProjectClick={handleProjectClick}
+      />
+
       {/* Footer */}
-      <footer className="footer-minimal enhanced-footer custom-footer">
-        <div className="container">
-          <div className="footer-content-minimal">
-            <div className="row align-items-center justify-content-between flex-nowrap">
-              <div className="col-auto d-flex align-items-center gap-3">
-                <img
-                  src="/images/img.jpeg"
-                  alt="Eduardo Wagner"
-                  className="footer-avatar"
-                  style={{
-                    width: "40px",
-                    height: "40px",
-                    borderRadius: "50%",
-                    objectFit: "cover",
-                    border: "2px solid #fff",
-                    boxShadow: "0 2px 12px rgba(30,136,229,0.10)",
-                  }}
-                />
-                <span
-                  style={{
-                    color: "#fff",
-                    fontWeight: 600,
-                    fontSize: "0.95rem",
-                  }}
-                >
-                  Eduardo Wagner &middot; Portfólio {new Date().getFullYear()}
-                </span>
-              </div>
-              <div className="col text-center d-none d-md-block">
-                <span
-                  className="footer-role"
-                  style={{
-                    color: "#b3d8f7",
-                    fontSize: "0.95rem",
-                    fontWeight: 500,
-                  }}
-                >
-                  Desenvolvedor Full Stack Júnior &nbsp;|&nbsp; Paraná, Brasil
-                </span>
-                <div style={{ color: "#b3d8f7", fontSize: "0.90rem" }}>
-                  "Transformando ideias em código, código em soluções, soluções
-                  em impacto"
-                </div>
-              </div>
-              <div className="col-auto text-end">
-                <div style={{ color: "#b3d8f7", fontSize: "0.80rem" }}>
-                  &copy; {new Date().getFullYear()} Todos os direitos
-                  reservados.
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </footer>
+      <Footer />
     </div>
   );
 }
 
-export default App
+export default App;
