+++ frontend/src/styles/index.css
@@ -196,6 +196,7 @@
   margin-bottom: 2rem;
   font-weight: 480;
   line-height: 1.6;
+  text-align: justify;
 }
 
 .floating-icon-4 {
